Remove budget and account nav, merge account settings
