+++ components/mobile-nav.tsx
@@ -53,16 +53,6 @@ export function MobileNav() {
               KGI/KPI
             </Link>
           </Button>
-          <Button variant="ghost" className="w-full justify-start gap-2">
-            <Home className="h-4 w-4" />
-            予算
-          </Button>
-          <Button asChild variant="ghost" className="w-full justify-start gap-2">
-            <Link href="/account">
-              <Wallet className="h-4 w-4" />
-              アカウント
-            </Link>
-          </Button>
           <Button asChild variant="ghost" className="w-full justify-start gap-2">
             <Link href="/settings">
               <Settings className="h-4 w-4" />
